refactor(srp): extract standup collectors to private module
standup_command.py mixed 4 responsibilities (CLI parsing, git/gh coleta,
LLM dispatch, Rich render). Extract parsing + coleta to _standup_collectors.py:
parse_since, parse_args, _ensure_git_repo, _ensure_gh_available,
collect_commits, collect_prs, collect_issues, collect_standup_data,
build_prompt, StandupData dataclass.

The command module keeps only StandupCommand class, generate_narrative
(LLM dispatch), and re-exports the moved symbols to preserve the
public surface that test_standup_command.py imports directly.
